move test_quantis test to test_tis
